# back/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "description": "API for managing users",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],

  "paths": {
    "/users/login": {
      "post": {
        "summary": "loggin users",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "loggin user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],

        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users/pix": {
      "post": {
        "summary": "Create a new pix",

        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pix"
              }
            }
          }
        },
        "responses": {
          
          "200": {
            "description": "The newly created pix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pix"
                }
              }
            }
          }
        }
      }
    },
    "/users/pix/{id}": {
      "get": {
        "summary": "Get pix for user id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "The Pix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pix"
                }
              }
            }
          }
        }
      }
    },
    "/users/pix/receive/{id}": {
      "get": {
        "summary": "Get receive pix for user id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "The Pix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pix"
                }
              }
            }
          }
        }
      }
    },
    "/users/pix/send/{id}": {
      "get": {
        "summary": "Get send pix for user id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "The Pix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pix"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "UnauthorizedError": {
        "description": "Acesso negado token não encontrado"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Pix": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "pix_key": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}